automationtestforbdd-develop/requirements.txt
behave==1.2.6
paver== 1.3.4
psutil==5.9.1
requests==2.28.1
pypiwin32==223
pyYAML==5.3.1
xlrd==2.0.1
allure-behave==2.9.45
Appium-Python-Client==1.0.2
selenium==3.141.0
Numpy==1.23.0
Pillow==9.1.1
pywinauto==0.6.8
pycryptodome==3.15.0
Paddlepaddle==2.3.1
wheel==0.37.1
Paddleocr==2.5.0.3